Allow full manipulation of Directives in AST node
This was one of the few (only?) exceptions to the AST being
full read-write, and prevented some processing of docs to
sanitize directives, for example, prior to execution.

There was really no apparent reason to introduce this limitation
at the AST level, so we implement `ICollection<Directive>` to
allow full mutation of this collection.